refactor: split terminal reliability modules
