+++ rust/alera-cli/src/terminal_host/session.rs
@@ -1,5 +1,4 @@
 use std::collections::HashSet;
-use std::io::{Read, Write};
 use std::sync::atomic::{AtomicU64, Ordering};
 use std::sync::mpsc::{sync_channel, SyncSender, TrySendError};
 use std::sync::Arc;
@@ -13,8 +12,12 @@ use crate::terminal_host::history_store::{TerminalHostCheckpoint, TerminalHostHi
 use crate::terminal_host::host_error::{HostError, HostResult};
 use crate::terminal_host::protocol::{encode_bytes, TerminalHostLaunch};
 
-/// Bytes read from the PTY per `read` call.
-const READ_CHUNK_BYTES: usize = 8192;
+mod io_threads;
+#[cfg(test)]
+mod tests;
+
+use io_threads::{spawn_reader, spawn_writer};
+
 const INPUT_QUEUE_CAPACITY: usize = 64;
 static NEXT_SESSION_INSTANCE_ID: AtomicU64 = AtomicU64::new(1);
 
@@ -487,303 +490,5 @@ impl Session {
             .upsert(checkpoint)
             .await
             .map_err(|error| HostError::state(error.to_string()))
-    }
-}
-
-/// Read the PTY on a dedicated thread, forwarding output and the final exit code
-/// through `on_event`.
-fn spawn_reader(
-    mut reader: Box<dyn Read + Send>,
-    mut child: Box<dyn portable_pty::Child + Send + Sync>,
-    on_event: Arc<dyn Fn(PtyEvent) + Send + Sync>,
-) {
-    std::thread::Builder::new()
-        .name("alera-pty-reader".to_string())
-        .spawn(move || {
-            let mut buffer = vec![0u8; READ_CHUNK_BYTES];
-            loop {
-                match reader.read(&mut buffer) {
-                    Ok(0) => break,
-                    Ok(read) => on_event(PtyEvent::Output(buffer[..read].to_vec())),
-                    Err(error) => {
-                        let code = child
-                            .try_wait()
-                            .ok()
-                            .flatten()
-                            .map(|status| status.exit_code() as i32)
-                            .unwrap_or(-1);
-                        on_event(PtyEvent::Error(error.to_string()));
-                        on_event(PtyEvent::Exit(code));
-                        return;
-                    }
-                }
-            }
-            let code = child
-                .wait()
-                .map(|status| status.exit_code() as i32)
-                .unwrap_or(-1);
-            on_event(PtyEvent::Exit(code));
-        })
-        .expect("failed to spawn pty reader thread");
-}
-
-fn spawn_writer(
-    mut writer: Box<dyn Write + Send>,
-    input_rx: std::sync::mpsc::Receiver<PtyWrite>,
-    on_event: Arc<dyn Fn(PtyEvent) + Send + Sync>,
-) {
-    std::thread::Builder::new()
-        .name("alera-pty-writer".to_string())
-        .spawn(move || {
-            let mut writer_failure: Option<String> = None;
-            while let Ok(request) = input_rx.recv() {
-                let error = match writer_failure.as_ref() {
-                    Some(message) => Some(message.clone()),
-                    None => writer
-                        .write_all(&request.bytes)
-                        .and_then(|_| writer.flush())
-                        .err()
-                        .map(|error| error.to_string()),
-                };
-                if writer_failure.is_none() {
-                    writer_failure.clone_from(&error);
-                }
-                on_event(PtyEvent::InputWritten {
-                    completion: request.completion,
-                    error,
-                });
-            }
-        })
-        .expect("failed to spawn PTY writer thread");
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    struct BlockingWriter {
-        started_tx: std::sync::mpsc::Sender<()>,
-        release_rx: std::sync::mpsc::Receiver<()>,
-    }
-
-    impl Write for BlockingWriter {
-        fn write(&mut self, bytes: &[u8]) -> std::io::Result<usize> {
-            self.started_tx.send(()).unwrap();
-            self.release_rx.recv().unwrap();
-            Ok(bytes.len())
-        }
-
-        fn flush(&mut self) -> std::io::Result<()> {
-            Ok(())
-        }
-    }
-
-    struct FailingWriter;
-
-    impl Write for FailingWriter {
-        fn write(&mut self, _bytes: &[u8]) -> std::io::Result<usize> {
-            Err(std::io::Error::other("writer failed"))
-        }
-
-        fn flush(&mut self) -> std::io::Result<()> {
-            Ok(())
-        }
-    }
-
-    fn test_session() -> Session {
-        Session {
-            instance_id: next_session_instance_id(),
-            id: "session-1".to_string(),
-            workspace_id: "workspace-1".to_string(),
-            tab_id: "tab-1".to_string(),
-            working_directory: "/repo".to_string(),
-            clients: HashSet::new(),
-            output_paused_clients: HashSet::new(),
-            buffer: ScrollbackBuffer::new(1024, &[]),
-            running: true,
-            exit_code: None,
-            ended_at: None,
-            master: None,
-            input_tx: None,
-            killer: None,
-            terminated: false,
-            checkpoint_gen: 0,
-            checkpoint_armed: false,
-            output_batch: Vec::new(),
-            output_batch_gen: 0,
-            output_batch_armed: false,
-            durable_output_batch: Vec::new(),
-            durable_output_batch_gen: 0,
-            durable_output_batch_armed: false,
-            durable_output_batch_sequence: 0,
-        }
-    }
-
-    #[test]
-    fn output_batch_coalesces_until_flush() {
-        let mut session = test_session();
-
-        assert_eq!(session.append_output(b"ab"), (Some(0), Some(0)));
-        assert_eq!(session.append_output(b"cd"), (None, None));
-        assert!(session.output_batch_due(0));
-
-        let batch = session.flush_output_batch().expect("batch");
-
-        assert_eq!(batch.payload["sessionId"], "session-1");
-        assert_eq!(
-            batch.payload["dataBase64"],
-            serde_json::Value::String(encode_bytes(b"abcd"))
-        );
-        assert_eq!(session.output_batch_len(), 0);
-        assert!(!session.output_batch_due(0));
-
-        let durable = session.flush_durable_output_batch().expect("durable batch");
-        assert_eq!(durable.data, b"abcd");
-        assert_eq!(durable.sequence, 0);
-    }
-
-    #[test]
-    fn output_batch_empty_flush_disarms_timer() {
-        let mut session = test_session();
-
-        assert_eq!(session.append_output(b"a"), (Some(0), Some(0)));
-        assert!(session.flush_output_batch().is_some());
-        assert!(session.flush_output_batch().is_none());
-
-        assert_eq!(session.append_output(b"b"), (Some(1), None));
-        assert!(!session.output_batch_due(0));
-        assert!(session.output_batch_due(1));
-        assert!(session.flush_output_batch().is_some());
-        let durable = session.flush_durable_output_batch().expect("durable batch");
-        assert_eq!(durable.data, b"ab");
-        assert_eq!(durable.sequence, 0);
-    }
-
-    #[test]
-    fn blocked_writer_applies_local_backpressure_without_blocking_sender() {
-        let (input_tx, input_rx) = sync_channel(1);
-        let (started_tx, started_rx) = std::sync::mpsc::channel();
-        let (release_tx, release_rx) = std::sync::mpsc::channel();
-        let (event_tx, event_rx) = std::sync::mpsc::channel();
-        let on_event: Arc<dyn Fn(PtyEvent) + Send + Sync> = Arc::new(move |event| {
-            event_tx.send(event).unwrap();
-        });
-        spawn_writer(
-            Box::new(BlockingWriter {
-                started_tx,
-                release_rx,
-            }),
-            input_rx,
-            on_event,
-        );
-
-        input_tx
-            .try_send(PtyWrite {
-                completion: PtyWriteCompletion::ClientRequest {
-                    client_id: 1,
-                    request_id: 10,
-                },
-                bytes: b"first".to_vec(),
-            })
-            .unwrap();
-        started_rx
-            .recv_timeout(std::time::Duration::from_secs(1))
-            .unwrap();
-        input_tx
-            .try_send(PtyWrite {
-                completion: PtyWriteCompletion::ClientRequest {
-                    client_id: 1,
-                    request_id: 11,
-                },
-                bytes: b"second".to_vec(),
-            })
-            .unwrap();
-        let full = input_tx.try_send(PtyWrite {
-            completion: PtyWriteCompletion::ClientRequest {
-                client_id: 1,
-                request_id: 12,
-            },
-            bytes: b"third".to_vec(),
-        });
-        assert!(matches!(full, Err(TrySendError::Full(_))));
-
-        release_tx.send(()).unwrap();
-        assert!(matches!(
-            event_rx
-                .recv_timeout(std::time::Duration::from_secs(1))
-                .unwrap(),
-            PtyEvent::InputWritten {
-                completion: PtyWriteCompletion::ClientRequest { request_id: 10, .. },
-                error: None,
-            }
-        ));
-        started_rx
-            .recv_timeout(std::time::Duration::from_secs(1))
-            .unwrap();
-        release_tx.send(()).unwrap();
-        assert!(matches!(
-            event_rx
-                .recv_timeout(std::time::Duration::from_secs(1))
-                .unwrap(),
-            PtyEvent::InputWritten {
-                completion: PtyWriteCompletion::ClientRequest { request_id: 11, .. },
-                error: None,
-            }
-        ));
-    }
-
-    #[test]
-    fn failed_writer_completes_every_queued_request_with_the_same_error() {
-        let (input_tx, input_rx) = sync_channel(3);
-        for request_id in 10..13 {
-            input_tx
-                .try_send(PtyWrite {
-                    completion: PtyWriteCompletion::ClientRequest {
-                        client_id: 1,
-                        request_id,
-                    },
-                    bytes: vec![request_id as u8],
-                })
-                .unwrap();
-        }
-        let (event_tx, event_rx) = std::sync::mpsc::channel();
-        let on_event: Arc<dyn Fn(PtyEvent) + Send + Sync> = Arc::new(move |event| {
-            event_tx.send(event).unwrap();
-        });
-
-        spawn_writer(Box::new(FailingWriter), input_rx, on_event);
-
-        for request_id in 10..13 {
-            assert!(matches!(
-                event_rx
-                    .recv_timeout(std::time::Duration::from_secs(1))
-                    .unwrap(),
-                PtyEvent::InputWritten {
-                    completion: PtyWriteCompletion::ClientRequest {
-                        request_id: actual,
-                        ..
-                    },
-                    error: Some(ref message),
-                } if actual == request_id && message == "writer failed"
-            ));
-        }
-    }
-
-    #[test]
-    fn exited_session_rejects_input_instead_of_leaving_request_pending() {
-        let mut session = test_session();
-        session.running = false;
-
-        let error = session
-            .queue_write(
-                PtyWriteCompletion::ClientRequest {
-                    client_id: 1,
-                    request_id: 10,
-                },
-                b"input",
-            )
-            .expect_err("exited session should reject input");
-
-        assert!(error.wire_message().contains("not running"));
     }
 }
