+++ rust/alera-core/src/runtime/orchestration_store_tests.rs
@@ -1,102 +1,9 @@
 use super::*;
 
-async fn store() -> (tempfile::TempDir, RuntimeStore) {
-    let dir = tempfile::tempdir().unwrap();
-    let store = RuntimeStore::open(dir.path()).await.unwrap();
-    (dir, store)
-}
+mod message_limits;
+mod support;
 
-fn message(
-    from: &str,
-    to: &str,
-    message_type: OrchestrationMessageType,
-) -> NewOrchestrationMessage {
-    NewOrchestrationMessage {
-        from_handle: from.to_string(),
-        to_handle: to.to_string(),
-        subject: "subject".to_string(),
-        body: "body".to_string(),
-        message_type,
-        priority: OrchestrationMessagePriority::Normal,
-        thread_id: None,
-        payload: None,
-    }
-}
-
-fn task(spec: &str, deps: Vec<String>) -> NewOrchestrationTask {
-    NewOrchestrationTask {
-        spec: spec.to_string(),
-        task_title: None,
-        display_name: None,
-        deps,
-        parent_id: None,
-        created_by_terminal_handle: None,
-    }
-}
-
-#[tokio::test]
-async fn message_limits_are_enforced_at_the_storage_boundary() {
-    let (_dir, store) = store().await;
-    let mut exact = message("a", "b", OrchestrationMessageType::Status);
-    exact.body = "x".repeat(ORCHESTRATION_BODY_MAX_BYTES);
-    store.insert_orchestration_message(exact).await.unwrap();
-
-    let mut oversized_body = message("a", "b", OrchestrationMessageType::Status);
-    oversized_body.body = "x".repeat(ORCHESTRATION_BODY_MAX_BYTES + 1);
-    assert!(store
-        .insert_orchestration_message(oversized_body)
-        .await
-        .unwrap_err()
-        .to_string()
-        .contains("body"));
-
-    let mut oversized_lifecycle = message("a", "b", OrchestrationMessageType::WorkerDone);
-    oversized_lifecycle.body = "x".repeat(ORCHESTRATION_LIFECYCLE_BODY_MAX_BYTES + 1);
-    assert!(store
-        .insert_orchestration_message(oversized_lifecycle)
-        .await
-        .unwrap_err()
-        .to_string()
-        .contains("body"));
-
-    let mut oversized_payload = message("a", "b", OrchestrationMessageType::Status);
-    oversized_payload.payload = Some("x".repeat(ORCHESTRATION_PAYLOAD_MAX_BYTES + 1));
-    assert!(store
-        .insert_orchestration_message(oversized_payload)
-        .await
-        .unwrap_err()
-        .to_string()
-        .contains("payload"));
-
-    let mut oversized_subject = message("a", "b", OrchestrationMessageType::Status);
-    oversized_subject.subject = "x".repeat(ORCHESTRATION_SUBJECT_MAX_BYTES + 1);
-    assert!(store
-        .insert_orchestration_message(oversized_subject)
-        .await
-        .unwrap_err()
-        .to_string()
-        .contains("subject"));
-
-    let mut oversized_thread = message("a", "b", OrchestrationMessageType::Status);
-    oversized_thread.thread_id = Some("x".repeat(ORCHESTRATION_THREAD_ID_MAX_BYTES + 1));
-    assert!(store
-        .insert_orchestration_message(oversized_thread)
-        .await
-        .unwrap_err()
-        .to_string()
-        .contains("thread id"));
-
-    let mut oversized_handle = message("a", "b", OrchestrationMessageType::Status);
-    oversized_handle.from_handle = "x".repeat(ORCHESTRATION_HANDLE_MAX_BYTES + 1);
-    assert!(store
-        .insert_orchestration_message(oversized_handle)
-        .await
-        .unwrap_err()
-        .to_string()
-        .contains("from handle"));
-
-    assert_eq!(store.orchestration_inbox(10).await.unwrap().len(), 1);
-}
+use support::{message, store, task};
 
 #[tokio::test]
 async fn read_and_delivered_are_independent() {
